refactor: Renaming and moving selected group context file
