+++ src/components/ContentResult.jsx
@@ -6,7 +6,7 @@ import { Content } from 'cozy-ui/transpiled/react/Layout'
 import { useI18n } from 'cozy-ui/transpiled/react/I18n'
 import useBreakpoints from 'cozy-ui/transpiled/react/hooks/useBreakpoints'
 
-import SelectedGroupContext from './Context'
+import SelectedGroupContext from './Contexts/SelectedGroup'
 import Header from './Header'
 import Toolbar from './Toolbar'
 import ContactsList from './ContactsList/ContactsList.jsx'
