+++ src/components/Contexts/SelectedGroup.jsx
@@ -1,7 +1,7 @@
 import React, { createContext, useState } from 'react'
 
 import { useI18n } from 'cozy-ui/transpiled/react/I18n'
-import { getDefaultSelectedGroup } from '../helpers/groups'
+import { getDefaultSelectedGroup } from '../../helpers/groups'
 
 const SelectedGroupContext = createContext()
